Change loss to Sparse Categorical Crossentropy
This is exactly the same as Categorical Crossentropy (what we used),
except that in the dataset we don't have to use one-hot encoding.
We can just use the token id and the loss will recognise that the
corresponding one-hot is correct. This just reduces the size of the
dataset in memory, especially as the token number increases.